backend/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.12.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
promise==2.3
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
